Move Button component from MissionsTable. Move Navbar from views to components.
