+++ src/components/Missions/MissionsButton.jsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { useDispatch } from 'react-redux';
+import { joinMission, leaveMission } from '../../redux/missions/missionsSlice';
+
+const MissionButton = ({ missionId, joinedMissions }) => {
+  const dispatch = useDispatch();
+
+  const handleJoinMission = () => {
+    dispatch(joinMission(missionId));
+  };
+
+  const handleLeaveMission = () => {
+    dispatch(leaveMission(missionId));
+  };
+
+  if (joinedMissions.includes(missionId)) {
+    return (
+      <button
+        className="leave-mission"
+        onClick={handleLeaveMission}
+        type="button"
+      >
+        Leave Mission
+      </button>
+    );
+  }
+  return (
+    <button
+      className="join-mission"
+      onClick={handleJoinMission}
+      type="button"
+    >
+      Join Mission
+    </button>
+  );
+};
+
+MissionButton.propTypes = {
+  missionId: PropTypes.string.isRequired,
+  joinedMissions: PropTypes.arrayOf(PropTypes.string).isRequired,
+};
+
+export default MissionButton;
