--- src/components/Missions/MissionsTable/MissionsTable.scss
@@ -1,118 +1,0 @@
-@import '../../variables';
-
-.missions {
-  text-align: left;
-  font-family: 'Roboto', sans-serif;
-  margin: 1rem 0 3rem;
-
-  table {
-    width: 100%;
-    border-collapse: collapse;
-    border: 1px solid $color-border;
-
-    .join-leave {
-      text-align: center;
-      padding: 0 5rem;
-    }
-
-    th,
-    td {
-      padding: 8px;
-      border: 1px solid $color-border;
-
-      &.mission-description {
-        font-family: 'Lato', sans-serif;
-        letter-spacing: 0.2px;
-        line-height: 1.6;
-      }
-
-      &.mission-name {
-        vertical-align: top;
-        font-weight: bold;
-      }
-
-      .member-status-badge {
-        display: inline-block;
-        padding: 4px 8px;
-        border-radius: 4px;
-        font-size: 14px;
-        font-weight: bold;
-      }
-
-      &.status-col {
-        pointer-events: none;
-        padding: 0.8rem;
-
-        .active-member {
-          background-color: $color-active-member;
-          font-weight: bold;
-          color: $color-text-light;
-          transition: 1s all ease-in-out;
-        }
-
-        .not-member {
-          background-color: $color-not-member;
-          text-transform: uppercase;
-          font-weight: bold;
-          color: $color-text-light;
-          transition: 1s all ease-in-out;
-        }
-      }
-
-      &.status-col,
-      &.mission-status {
-        text-align: center;
-        white-space: nowrap;
-        margin: 0.2rem;
-      }
-
-      &.mission-status {
-        vertical-align: middle;
-
-        button {
-          font-family: 'Roboto', sans-serif;
-          background: transparent;
-          padding: 0.5rem;
-          cursor: pointer;
-
-          &.join-mission {
-            color: $color-text-dark;
-            border: 2px solid $color-btn-join;
-            transition: 1s all ease-in-out;
-
-            &:hover {
-              background-color: $color-text-dark;
-              color: $color-text-light;
-              transition: all 0.3s ease-in;
-            }
-          }
-
-          &.leave-mission {
-            border: 2px solid $color-leave-mission;
-            color: $color-leave-mission;
-            transition: 1s all ease-in-out;
-
-            &:hover {
-              background-color: $color-leave-mission;
-              color: $color-text-light;
-              transition: all 0.3s ease-in;
-            }
-          }
-        }
-      }
-
-      &.mission-status > button {
-        overflow: hidden;
-        white-space: nowrap;
-      }
-
-      th {
-        font-weight: bold;
-      }
-
-      td:last-child {
-        text-align: center;
-      }
-    }
-  }
-}
